Revert "rm ghoust_test.py - no idea what that was used for"
This reverts commit 59a0c12ead31be365cb96024c4b45c4cb1fa1554.

I "undo" this dick move of removing the base for actual
unit tests.

Looking at the commit I get the bewildering feeling of
someone who goes into a datacenter and turns of running
servers with the same "no idea what that was used for"
reasoning.

The sad thing about that is, that there are stories out
there of DevOps/Administrators who knew someone who actually
did things like that.